move ScoredPointTies to segment, make inner by reference
