+++ lib/segment/src/utils/mod.rs
@@ -2,4 +2,5 @@ pub mod fmt;
 pub mod fs;
 pub mod mem;
 pub mod path;
+pub mod scored_point_ties;
 pub mod tar;
