+++ lib/segment/tests/integration/payload_index_test.rs
@@ -34,10 +34,9 @@ use segment::types::{
     PayloadSchemaParams, PayloadSchemaType, Range, SegmentConfig, VectorDataConfig,
     VectorStorageType, WithPayload,
 };
+use segment::utils::scored_point_ties::ScoredPointTies;
 use serde_json::json;
 use tempfile::Builder;
-
-use crate::utils::scored_point_ties::ScoredPointTies;
 
 const DIM: usize = 5;
 const ATTEMPTS: usize = 100;
@@ -639,17 +638,17 @@ fn test_struct_payload_index() {
 
         // Perform additional sort to break ties by score
         let mut plain_result_sorted_ties: Vec<ScoredPointTies> =
-            plain_result.iter().map(|x| x.clone().into()).collect_vec();
+            plain_result.iter().map(|x| x.into()).collect_vec();
         plain_result_sorted_ties.sort();
 
         let mut struct_result_sorted_ties: Vec<ScoredPointTies> =
-            struct_result.iter().map(|x| x.clone().into()).collect_vec();
+            struct_result.iter().map(|x| x.into()).collect_vec();
         struct_result_sorted_ties.sort();
 
         plain_result_sorted_ties
             .into_iter()
             .zip(struct_result_sorted_ties.into_iter())
-            .map(|(r1, r2)| (r1.scored_point, r2.scored_point))
+            .map(|(r1, r2)| (r1.0, r2.0))
             .for_each(|(r1, r2)| {
                 assert_eq!(r1.id, r2.id, "got different ScoredPoint {r1:?} and {r2:?} for\nquery vector {query_vector:?}\nquery filter {query_filter:?}\nplain result {plain_result:?}\nstruct result{struct_result:?}");
                 assert!((r1.score - r2.score) < 0.0001)
