+++ lib/segment/tests/integration/utils/mod.rs
@@ -1,1 +1,1 @@
-pub mod scored_point_ties;
+
